# api/docs/swagger.yaml
definitions:
...
  ek_entity_service.EntityFilesSwag:
    properties:
      comment:
        type: string
    type: object
...
  ek_setting_service.RegionFilesSwag:
    properties:
      comment:
        type: string
      file_name:
        type: string
      region_id:
        type: string
      type:
        type: string
      url:
        type: string
    type: object
...
  ek_variables.FailureResponse:
    properties:
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  v1.FilesResponse:
    properties:
      file_path:
        type: string
      file_url:
        type: string
    type: object
info:
  contact: {}
  license: {}
paths:
...
  /v1/file-upload:
    post:
      consumes:
      - multipart/form-data
      - application/json
      description: API for creating file
      parameters:
      - description: file
        in: formData
        name: file
        required: true
        type: file
      - description: region
        in: body
        name: region
        required: true
        schema:
          $ref: '#/definitions/ek_entity_service.EntityFilesSwag'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.FilesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
      summary: Create file
      tags:
      - file_upload
...
  /v1/image-upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: image
        in: formData
        name: image
        required: true
        type: file
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.FilesResponse'
      tags:
      - file_upload
...
  /v1/region/{region_id}/upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: file
        in: formData
        name: file
        required: true
        type: file
      - description: region_id
        in: path
        name: region_id
        required: true
        type: string
      - description: body
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/ek_setting_service.RegionFilesSwag'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.FilesResponse'
      tags:
      - file_upload
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"